refactor(ui): extract feed/observation-view.ts — view-mode selection helpers
